requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.10
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
Brotli==1.1.0
brotlicffi==1.1.0.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cramjam==2.8.3
cryptography==43.0.1
cycler==0.12.1
fastparquet==2024.5.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.3.0
future==1.0.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
grpcio==1.67.0
gviz-api==1.10.0
huggingface_hub==0.26.5
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lz4==4.3.3
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
pillow==11.0.0
protobuf==5.28.3
psutil==6.0.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycares==4.4.0
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.4
pyspnego==0.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.1.post0
six==1.16.0
smbprotocol==1.14.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.11.1
tensorboard-plugin-wit==1.7.0
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1a0
tqdm==4.66.5
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wheel==0.44.0
yarl==1.13.1
